+++ src/test/java/org/cicirello/math/rand/EnhancedRandomGeneratorTests.java
@@ -185,229 +185,6 @@ public class EnhancedRandomGeneratorTests {
 	}
 	
 	@Test
-	public void testBinomials() {
-		class Wrapper {
-			boolean different;
-			int last = -1;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.binomials(100, 0.5).limit(10).forEach(
-			x -> {
-				assertTrue(x < 100);
-				assertTrue(x > 0);
-				if (w.last >= 0 && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-			}
-		);
-		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testBinomialsLimited() {
-		class Wrapper {
-			boolean different;
-			int last = -1;
-			int count;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.binomials(5L, 100, 0.5).forEach(
-			x -> {
-				assertTrue(x < 100);
-				assertTrue(x > 0);
-				if (w.last >= 0 && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-				w.count++;
-			}
-		);
-		assertTrue(w.different);
-		assertEquals(5, w.count);
-	}
-	
-	@Test
-	public void testExponentials() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.exponentials().limit(10).forEach(
-			x -> {
-				assertTrue(x >= 0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-			}
-		);
-		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testExponentialsLimited() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-			int count;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.exponentials(5L).forEach(
-			x -> {
-				assertTrue(x >= 0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-				w.count++;
-			}
-		);
-		assertTrue(w.different);
-		assertEquals(5, w.count);
-	}
-	
-	@Test
-	public void testCauchys() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.cauchys(500.0, 0.01).limit(10).forEach(
-			x -> {
-				assertTrue(x < 501.0);
-				assertTrue(x > 499.0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-			}
-		);
-		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testCauchysLimited() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-			int count;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.cauchys(5L, 500.0, 0.01).forEach(
-			x -> {
-				assertTrue(x < 501.0);
-				assertTrue(x > 499.0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-				w.count++;
-			}
-		);
-		assertTrue(w.different);
-		assertEquals(5, w.count);
-	}
-	
-	@Test
-	public void testGaussians() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.gaussians().limit(10).forEach(
-			x -> {
-				assertTrue(x < 5.0);
-				assertTrue(x > -5.0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-			}
-		);
-		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testGaussiansMuSigma() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.gaussians(500.0, 0.01).limit(10).forEach(
-			x -> {
-				assertTrue(x < 500.05);
-				assertTrue(x > 499.95);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-			}
-		);
-		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testGaussiansMuSigmaLimited() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-			int count;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.gaussians(5L, 500.0, 0.01).forEach(
-			x -> {
-				assertTrue(x < 500.05);
-				assertTrue(x > 499.95);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-				w.count++;
-			}
-		);
-		assertTrue(w.different);
-		assertEquals(5, w.count);
-	}
-	
-	@Test
-	public void testGaussiansLimited() {
-		class Wrapper {
-			boolean different;
-			double last = Double.NaN;
-			int count;
-		}
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-		final Wrapper w = new Wrapper();
-		erg.gaussians(5L).forEach(
-			x -> {
-				assertTrue(x < 5.0);
-				assertTrue(x > -5.0);
-				if (w.last != Double.NaN && w.last != x) {
-					w.different = true;
-				}
-				w.last = x;
-				w.count++;
-			}
-		);
-		assertTrue(w.different);
-		assertEquals(5, w.count);
-	}
-	
-	@Test
 	public void testIntsOriginBound() {
 		class Wrapper {
 			boolean different;
@@ -495,123 +272,6 @@ public class EnhancedRandomGeneratorTests {
 		);
 		assertEquals(5, w.count);
 		assertTrue(w.different);
-	}
-	
-	@Test
-	public void testBinomial() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		for (int i = 0; i < 10; i++) {
-			int lowB = erg.nextBinomial(100, 0.05);
-			assertTrue(lowB < 100);
-			int highB = erg.nextBinomial(100, 0.95);
-			assertTrue(highB > 0);
-			int midB = erg.nextBinomial(100, 0.5);
-			assertTrue(midB < 100);
-			assertTrue(midB > 0);
-		}
-	}
-	
-	@Test
-	public void testCauchyScale() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int positive = 0;
-		int negative = 0;
-		double scale = 0.5;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextCauchy(scale);
-			if (g > 0) positive++;
-			if (g < 0) negative++;
-			assertTrue(Math.abs(g) < 100.0*scale);
-		}
-		assertTrue(positive > 0);
-		assertTrue(negative > 0);
-	}
-	
-	@Test
-	public void testCauchyMedianScale() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int greater = 0;
-		int lesser = 0;
-		double scale = 0.05;
-		double median = 42.0;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextCauchy(median, scale);
-			if (g > median) greater++;
-			if (g < median) lesser++;
-			assertTrue(Math.abs(g-median) < 100.0*scale);
-		}
-		assertTrue(greater > 0);
-		assertTrue(lesser > 0);
-		median = -42.0;
-		greater = 0;
-		lesser = 0;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextCauchy(median, scale);
-			if (g > median) greater++;
-			if (g < median) lesser++;
-			assertTrue(Math.abs(g-median) < 100.0*scale);
-		}
-		assertTrue(greater > 0);
-		assertTrue(lesser > 0);
-	}
-	
-	@Test
-	public void testGaussian() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int positive = 0;
-		int negative = 0;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextGaussian();
-			if (g > 0) positive++;
-			if (g < 0) negative++;
-			assertTrue(Math.abs(g) < 5.0);
-		}
-		assertTrue(positive > 0);
-		assertTrue(negative > 0);
-	}
-	
-	@Test
-	public void testGaussianStDev() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int positive = 0;
-		int negative = 0;
-		double stdev = 0.05;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextGaussian(stdev);
-			if (g > 0) positive++;
-			if (g < 0) negative++;
-			assertTrue(Math.abs(g) < 5.0*stdev);
-		}
-		assertTrue(positive > 0);
-		assertTrue(negative > 0);
-	}
-	
-	@Test
-	public void testGaussianMeanStDev() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int greater = 0;
-		int lesser = 0;
-		double stdev = 0.05;
-		double mean = 42.0;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextGaussian(mean, stdev);
-			if (g > mean) greater++;
-			if (g < mean) lesser++;
-			assertTrue(Math.abs(g-mean) < 5.0*stdev);
-		}
-		assertTrue(greater > 0);
-		assertTrue(lesser > 0);
-		mean = -42.0;
-		greater = 0;
-		lesser = 0;
-		for (int i = 0; i < 20; i++) {
-			double g = erg.nextGaussian(mean, stdev);
-			if (g > mean) greater++;
-			if (g < mean) lesser++;
-			assertTrue(Math.abs(g-mean) < 5.0*stdev);
-		}
-		assertTrue(greater > 0);
-		assertTrue(lesser > 0);
 	}
 	
 	@Test
